+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - .
 
 ### Changed
-- .
+- (**Extension**) Support fully qualified extension class names
 
 ### Fixed
 - .